localizations helper does not crash when rendering a newly saved widget (although it does not actually offer localizations until page refresh due to a deeper apostrophe issue). Also, merged helper implementations into helpers.js.
