+++ lib/api.js
@@ -847,20 +847,6 @@ module.exports = function(self, options) {
     });
   };
 
-  self.enableHelpers = function() {
-    self.addHelpers({
-      localizations: function() {
-        var localizations = [];
-        _.each(self.apos.templates.contextReq.data.workflow.localizations, function(localization, locale) {
-          if (!self.locales[locale].private) {
-            localizations.push(localization);
-          }
-        });
-        return localizations;
-      }
-    });
-  };
-
   // Returns the same data structure as `self.nestedLocales`,
   // except that any locales that are not editable by the
   // user indicated by `req` (that is, the user does not have
