+++ lib/callAll.js
@@ -188,12 +188,9 @@ module.exports = function(self, options) {
       self.apos.templates.addBodyClass(req, 'apos-workflow-live-page');
     }
 
-    // Pass on workflow-related information to Nunjucks templates,
-    // notably `data.workflow.context` which will be the page or piece
+    // Pass on `data.workflow.context` which will be the page or piece
     // the user thinks of as the "context" for the current page rendering
 
-    req.data.workflow = req.data.workflow || {};
-    _.assign(req.data.workflow, _.pick(self, 'locale', 'nestedLocales'));
     var context = self.getContext(req);
     if (context && context.workflowGuid) {
       req.data.workflow.context = context;
