+++ lib/middleware.js
@@ -15,6 +15,9 @@ module.exports = function(self, options) {
     middleware: function(req, res, next) {
       var to, host, matches, locale, prefix, hostname, subdomain, domain, candidates;
 
+      req.data.workflow = req.data.workflow || {};
+      _.assign(req.data.workflow, _.pick(self, 'locales', 'nestedLocales'));
+
       // Hack to implement bc with the subdomains option
       // by populating self.hostnames for each prefix, as soon as we
       // actually know what the domain name is
